Moving configuration files to the Confs folder Adding async / await and disabling Selenium control flow manager
